Add `isdag` method to determine if a dask is a DAG, and add non-recursive `get`.
Both of these need better documentation and tests, but I wanted to share.

`isdag` is a pretty standard DFS, which is stack-based (referring to data
container), not recursive.

The non-recursive `get` is only used when the "get" keyword isn't given.
It can handle arbitrarily large graphs, because it is not limited by
Python's recursion depth limit.  It is also faster than the recursive method.