Async scheduler uses FIFO policy in task selection
Previously we inspected the current dependencies to find tasks
that would allow us to release data.  Our implementation for this was
linear in the size of the ready-to-run tasks.  For large blocked
algorithms this can become a bottleneck.

Now we use a first-in-first-out policy so that we run newly ready tasks
instead of long lingering ones.  We hope that this causes somewhat
independent chains to execute before starting new ones.