da.store uses duck typing and no longer has a return value
It's not very friendly (or expected) to use isinstance checks for list/tuple
in a function that claims it supports sequences.

I also didn't see a point to giving this function a return value when it
explicitly modifies the data in-place. Note also that I'm pretty sure the
`single_output` logic was broken in the existing function.