add dask.array.cache
This allows us to evaluate a dask.array, store in some cache,
and then return a new dask array pointing to that precomputed data.

One can specify a cache as either a MutableMapping or as an ndarray

Example
-------

>>> import dask.array as da
>>> x = da.arange(5, chunks=2)
>>> y = 2*x + 1
>>> z = y.cache()  # triggers computation

>>> y.compute()  # Does entire computation
array([1, 3, 5, 7, 9])

>>> z.compute()  # Just pulls from store
array([1, 3, 5, 7, 9])

You might base a cache off of an array like a numpy array or
h5py.Dataset.

>>> cache = np.empty(5, dtype=x.dtype)
>>> z = y.cache(store=cache)
>>> cache
array([1, 3, 5, 7, 9])

Or one might use a MutableMapping like a dict or chest

>>> cache = dict()
>>> z = y.cache(store=cache)
>>> cache  # doctest: +SKIP
{('x', 0): array([1, 3]),
 ('x', 1): array([5, 7]),
 ('x', 2): array([9])}

Fixes #293