WIP: dask.array.store_context
Posting this for feedback -- I know it needs a bit more polish.

This function stores dask arrays in array-like objects created and
cleaned up by context-managers.

This is useful, for example, if you would like to store arrays in a very
larger number of files, such that creating each of the files in which to
store these arrays ahead of time would exhaust your file pointers.

Example usage:

    >>> import h5py
    >>> @contextlib.contextmanager
    ... def out(n, shape, dtype):
    ...     with h5py.File('%s.hdf5' % n) as f:
    ...         yield f.create_dataset('data', shape=shape, dtype=dtype)

    >>> target_contexts = [out(n, a.shape, a.dtype) for n, a in enumerate(arrays)]
    >>> store_context(arrays, target_contexts)