Add an rewrite rule for handling (getitem, (elemwise, x), i)
These should be rewritten from `(getitem, (elemwise, x), i)` to
`(elemwise, (getitem, x, i))`. In many cases, this translates into a
significant reduction in IO.

This is particularly relevant for computed coordinates, which are ubiquitous
in the fields of meteorology and occeanography. These case typically be
calculated in a straightforward way from ufuncs, and we would like to support
these using dask in xray (see https://github.com/xray/xray/issues/462).