Add `pure` kwarg to `do`
Can now specify if a wrapped function is pure or not (default is False).
This is important for functions like `random` or `time`.

Also changed `tokenize` to default to incremental token if hashing
fails. This means we don't do common-subexpression-elimination as
aggressively as we did earlier, but people shouldn't be relying on dask
preventing repeat computations anyway. This is probably a more sane
default than we had originally.