Make progress bar work in notebook
Call to sys.stdout.flush() in posttask is needed because ipython buffers
printed output from background threads in the notebook. This gets around
that by calling flush in the main thread. This means that the
progressbar will not update as frequently in the notebook as it would in
the terminal, but still works in both.