Use new tokenize method in dask
Tokenize is now variadic, and maps normalize over all arguments.
Dispatch is used for `normalize` to allow extension to custom types. The
following rule of thumb is used when applying `tokenize`:

- If special casing is needed (sorting kwargs, handling pandas objects,
  etc...), pass in each argument separately.
- Else, pass in a tuple containing all arguments.

This allows the dispatch system to handle the special cases, while also
remaining fast for normal use or cases that were already handled locally in each
function.