Automatic rechunking and coercion of numpy.ndarrays
Fixes GH290

The implementation of rechunking here is the most basic version possible: it
only succeeds if at most one of the arrays has multiple chunks defined along
an axis.

This suffices for coercing ndarrays -- in the future we can consider more
complete solutions.

Also includes a change to rechunking, where it doesn't bother if the chunk
size is unchanged. We could remove this (and do it in unify_chunks instead),
but it seemed like a general improvement.