propagate tokens down from grid_search to pipeline
Previously every call to pipeline.fit/score would tokenize the input
arrays.  This can be costly, particularly when the inputs are text.

Now we tokenize once in grid_search and pass these results down to
pipeline.  This improves efficiency but increases bookkeeping and makes
it harder to separate sklearn from dask code.

As an alternative we might consider temporarily caching `tokenize` by
id.