use sentinel rather than nbytes header in read/write
Having two `yield read(...)` statements back to back resulted
in a lot of latency (100ms) as we go through the other waiting
coroutines.

Now we use sentinels which are supported directly in tornado.
Latency goes way down as a result.  Unfortunately the raw sockets used
in the *_sync functions don't support sentinels by default, so we have
to add this logic ourselves.  Our solution is simple, but very
inefficient.