Expand tensordot to expose intermediates
Previously the main operation used to compute da.tensordot computed
the chunkwise tensordots and summed them together in a single in-memory
operation.  This possibly resulted in memory blowup, particularly in
short-and-fat .dot. tall-and-skinny cases.

Now we treat each sub-tensordot call as an independent task.  This
allows the scheduler to clear out intermediate results more
intelligently.  We expose more of the algorithm to the scheduler.

Fixes #821