Finalize no longer accepts the dask object
This was tricky when putting finalize into the dask graph itself,
as is necessary when we ask a remote worker to compute a dask object.

In these cases we had a full copy of the dask object within the dask
graph, which caused some serialization issues.

Now the _finalize staticmethods of each collection only accept the
results and don't have access to any of the object metadata.