Add cancel function to executor
Cancel accepts a list of futures and tries to stop their execution on the
cluster.  It removes the stated desire that this client wants these keys, and
all dependent keys to be computed.  If no other client also wants these keys
then a general sweep through distributed memory will occur.  This will not stop
computations that are currently in progress however (it is quite hard to
interrupt a thread)

Examples
--------

>>> e.cancel([a, b, c])

>>> future.cancel()

Fixes #111