Add ready deque to scheduler
Previously we assigned a task to a worker once that task was ready to
run.  It would then linger in that worker's stack, possibly for a long
time.  If in the future another worker capable of running that task
became free we had no mechanism for it to steal work.  There was good
reason for this, it's tricky to find good tasks to steal in constant
time.

However, if the common case of leaf tasks that have neither dependent
tasks nor restrictions it makes sense to keep these ready tasks in a
common pool from which all workers can draw.  This is the new `ready`
collection on the scheduler, stored as a FIFO deque.

This enables workers to load balance more effectivley in this common
case and allows newly arrived workers a pool of available work.