Fix bug in ResourceProfiler resource cleanup
Previously a `__del__` method was on the `_Tracker` class, which would
join the tracker upon cleanup. This worked fine if the parent process gc
hit it first, but failed if the child process gc did (as child processes
can't join themselves). After this happens, if a new resource profiler
was created, then it would try to grab information from all of its
sibling processes, including the one that failed to cleanup properly.
Process information can't be gathered from a zombie process, so this
would error, leading to another zombie process, and so on...

The fix is to move the `__del__` cleanup to the `ResourceProfiler`
class, and also filter out zombie processes in the `_Tracker` class.

Fixes #1094.