Adding new callback for the start that sees the state.
The `start(dsk)` callback occurs before the state is initialized,
so that it can mess with the dask before the state is setup.  It
could also be useful to have a callback before task execution
begins but *after* the state is setup so it can modify the state
itself.  I want to do this for restarting dask computations in
callbacks.