Pass dependencies through optimizations
This makes cull and fuse, two very common optimizations, return a
second ``dependencies`` result that is a frequently reused side effect
of the computation.  Dask.array was recreating this fairly expensive
data five times for a single optimization pass.

This is bad because it complicates optimizations.  This is good
because it halves total optimization time in some cases.