Don't inline getitem in dask.array
Previously we would inline getitem, getarray, and transpose calls
because they were very cheap to recompute.  When we start considering
distributed memory systems this optimization ends up causing a lot more
data transfer than is necessary.

However, it's still very important to inline the initial getitem call
coming from an HDF5 or NetCDF file.  This avoids sucking a lot of data
into memory in common operation.  We still want to inline in this case.

As a compromise we inline getarray but not getitem and use getarray
when dealing with foreign data sources that may or may not provide
numpy arrays on slicing.