Add lineterminator to read_csv keywords
Previously we avoided putting lineterminator into keywords because we
often need to split on `\r\n` but can't pass this into pandas (only
terminators of length one are supported).  This caused problems when
the user explicitly stated that they were using `\n`, which has slightly
different behavior from default.

Now we pass in the lineterminator to pandas if it is of length one and
ignore it in pandas otherwise (staying with default).

Long term we might consider breaking this up into two arguments, one
for block splitting and one for within-block parsing.

Fixes #1132