Inline getattr for dask.delayed objects
When using methods with `Delayed` values, a task is first created for
the attribute access, and a second task is created for the method call.
While the scheduler overhead of doing two tasks is minimal, this causes
problems when using delayed with distributed/multiprocessing, as methods
aren't necessarily pickleable. To fix this, we inline all calls to
`getattr` (which should also improve performance, since these calls are
cheap).