Ensure to_task_dasks always returns a task
Previously lists were special cased to return a list of keys. This
worked fine, except when fusing was used. Fusing resulted in a list
of tasks, which the scheduler wouldn't recognize as a task unless there
were external dependencies involved.

This fix ensures that any task returned by `to_task_dasks` is an actual
task. Nested lists can still contain lists, as long as the outer thing
is a task.