Fix bug in slicing optimization
When slicing an unknown array-like object, ``getarray`` is used to
ensure that the output of the slicing operation is a numpy array,
instead of another array-like. When optimizing nested slices, we need to
ensure that if ``getarray`` is used at all in the chain, then the
slicing still happens. A bug in this logic was introduced in the recent
work on speeding up the optimization passes. This patch fixes this bug,
and reverts the changes to the tests made that allowed it to pass.