Fix scheduler behavior for top-level lists
Previously lists were treated differently depending on if they were a
toplevel key in the graph or not. The new behavior treats them the same
as long as they're not nested lists. This is best illustrated with an
example:

```python
dsk = {'a': [1, 2, 3],
       'b': 'a',
       'c': [(inc, 1), 2],
       'd': [[1, 2, (inc, 1)], 4],
       'e': [(list, [1, 2, (inc, 1)]), 4],
       'f': ['a', 'b']}
```

In this example:
- 'a' is data, as none of the elements are keys or tasks
- 'b' is not data, it's just an alias to 'a' (a trivial task)
- 'c' is not data, as it contains a task
- 'd' is data, as the list contains no tasks.
- 'e' is not data, as it contains tasks. Note that the result of this is
  ``[[1, 2, 3], 4]``, which is probably the desired result for `d`.
- 'f' is not data as it contains keys in the graph

To summarize:

A value in the graph is computed if:
- It's a task
- It's a key in the graph
- It's a list that contains either tasks or keys