Shard large frames in protocol
Previously the protocol would fail on large values, such as might occur
when sending around 4GB numpy arrays.  While this is generally
discouraged it would be nice to avoid a complete system halt.

Now we identify these large bytestrings and shard them accordingly,
compressing each part as an individual chunk.  This incurs an extra
memory copy (we're not being overly careful) but if effective.

Fixes #321