Multiprocess scheduler handles unpickling errors
Previously if a task failed to be cleanly deserialized, the error wasn't
handled gracefully and it would cause the scheduler to lock up. This
fixes that issue by special casing the `apply_async` implementation in
`dask.multiprocessing`. This means that it no longer matches the call
signature of `pool.apply_async`, but in reality we always call
`apply_async` with the same arguments. The special casing allows for
deserialization errors to be handled the same as all other errors, and
allows for a clean exit.

Also did a few small pep8 cleanups in the touched files.