Ignore index when concating trivial indices
If all partitions have trivial indices (unnamed `RangeIndex` starting at
0 and incrementing by 1), we now concatenate together while ignoring the
indices. Note that this isn't a perfect solution, as filtering steps
might get rid of the trivial index (e.g. `df.a[df.a > 10]` will convert
a `RangeIndex` to an `Int64Index`. In those cases we fall back to the
current behavior of just concatenating the indices.

Fixes #1047.