Unify shuffle around rearrange_column
This unifies shuffles and set_index around a single standard approach.
We create a new column that dictates the target partition for each row
and then depend on either partd or task-shuffling to rearrange into
rows.

Various functions that use shuffling create this column and consume the
resulting dataframe in different ways, but they all share this core