Break ordering ties by str
This changes task ordering to break ties using the string representation
of the key.  This is particularly useful when computing dask collections
on an asynchronous scheduler, because the beginning parts of the
computation start first.  This will allow for `df.head()` for example to
work almost immediately after `e.persist(df)`.  The keys like `('df',
0)` will run before keys like `('df', 242)`.

I haven't benchmarked this to see if it significantly affects overhead.