Protocol supports large bytes in nested objects (#397)
* Protocol supports large bytes in nested objects

Previously the distributed protocol would only shard large bytestrings
if they were values in the top level dictionary. This would cause
problems if there were large bytestrings in nested objects.

To overcome this, we now shard all large bytestrings if they meet the
following requirements:

- The top level of the message is a dictionary
- They are contained in any level of `dict`/`list` nesting. Other
  objects aren't traversed.

Fixes #393.

* dumps protocol change no longer mutates msg

We now do a `deepcopy` if any of the values were determined to be large,
and mutate the copy instead of the original. This should be cheap in the
common case where no copy is needed.

* Change extract and shard constants

Protocol now uses the following constants
- Extract at 2**20 bytes
- Shard at 2**28 bytes

Also improved tests a bit.