Refactor Scheduler (#398)
This is a large refactor of the distributed scheduler.

### Problem

* Recently we've found deadlocks in the scheduler.  These have become increasingly time consuming to identify and resolve.
* No one feels comfortable operating on the scheduler code except for myself

Historically as we've added more and more features we convolute the scheduler code more and more.  This has led to a difficult and time consuming situation where resilience and performance are hard to guarantee, and hard to fix once an issue has been identified.

The previous architecture was a set of scheduler methods that called each other in a difficult-to-reason-about way that were triggered by a few events.

### Solution

This PR adds a bit of top-down design to the management of scheduler state.  This has a few positive effects:

1.  The resulting code is cleaner and easier to follow
2.  The design of the scheduler is easier to talk about.  What were previously conventions and ways of thinking about what's happening have become actual function calls
3.  Much of the state changes happen in a central place.  This enables consistent logging, and runtime checking, which greatly accelerate the debugging and development process.

### Transitions and Stimuli

To learn more about the new design you should read the doc page that comes along with this PR

* Docs: https://github.com/mrocklin/distributed/blob/scheduler-refactor/docs/source/scheduler.rst
* Code: https://github.com/mrocklin/distributed/blob/scheduler-refactor/distributed/scheduler.py