Fix pickle for all classes (#1457)
* Fix pickle for all classes

Not all dask collections were pickleable. Further, the dataframe
implementation was broken, resulting in a shared `__dict__` between
objects. This fixes that, and adds tests for all classes to ensure
pickleable.

* Fix bag pickle on py3.3

Use __getstate__ and __setstate__ instead. This required changing a
`str` to a property. To forward the docstring, `StringAccessor` had to
be moved before `Bag` in the file.