Resolve intermittent failures (#415)
* improve logging for worker messages

* Reduce size of dataframes

A recent commit expanded the size of the dataframes here in this test,
causing timeout errors.

It takes a long time because we're creating n**2 groups

* allow many_progress test to be a little bit late

* Fix cancellation error

When cancelling keys from update_graph the keys aren't yet set in
dependents.  So we allow this absence in client_releases_keys.

* stress_creation_and_deletion robust to living nanny

Sometimes the nanny doesn't die

* Fix when moving to processing just before worker fails

Transition_ready_processing sends a task to a worker.  This can fail
if the worker dies just at the right time.  We've correctly been
calling remove_worker in this context, however we did this in the
middle of the transition, which was an unsafe state.  Now we do
it at the end