Rewrite dask.array.reshape
It now handles more cases efficiently, including the one identified in GH1468. I
*think* I've now covered every scenario where it's possible to reshape in
contiguous order by calling `np.reshape` on each chunk, but I haven't proved it.

In cases where we cannot reshape without creating a ridiculously large number of
tasks, it now fails noisily with a ValueError.

This is a breaking change, but one that I think will result in a better user
experience. Let's save it for a major dask release (0.11?).