Refactor `Scalar` class
Previously the `Scalar` class wouldn't track metadata at all. This
caused problems with metadata inference if `Scalar` was used in any
operation. To fix this, the `Scalar` class has been refactored as
follows:

- Each scalar has a `_meta` attribute, same as all other
  `dask.dataframe` objects. This attribute is a non-empty scalar of the
  appropriate type. Supported types include any subclass of
  `np.generic`, as well as `str`/`unicode`, and custom pandas scalars
  like `pd.Timestamp` and `pd.Timedelta`. A corresponding
  `_meta_nonempty` attribute is also provided - this just returns
  `_meta`.

- Inference is added to `reduction` and `Scalar` returning
  `map_partitions`. Operators that work on scalars are modified to infer
  metadata based off the scalar dtype as well.

- `make_meta` is modified to dispatch on raw `dtype` arguments. This
  means that `make_meta('i8')` will produce the appropriate metadata for
  a scalar containing an int64.

Caveat: If a type is supported as a numpy scalar, a numpy scalar is
used. This does mean that for some (rare) operations, the `_meta` object
on a scalar may be a `np.int64`, but the computed result is an `int`.