Make align_partitions more strict (#1475)
Previously when aligning DataFrame/Series with unknown divisions, and
ones with known divisions, bad things would occur. Now we explicitly
check if any of the input objects have unknown divisions, and fail if
they do.

A separate function used elsewhere is `_maybe_align_partitions` - this
only calls `align_partitions` if the divisions are unequal. This means
that *equal but unknown* divisions can propogate through, which is
desired for many elementwise operations. The case of mixing known and
unknown is then caught in `align_partitions`, failing as desired.

Tests for these behaviors are then added to ensure the fail cases hold.

Also Added tests for known and unknown divisions for `assign` (testing
that this fixes #1373).