Create new threadpool when operating from thread (#1487)
* Create new threadpool when operating from thread

When we call dask.threaded.get from a separate thread we shouldn't use
the default threadpool.  ThreadPool objects are not safe in this manner.

Now, if we call dask.threaded.get from the non-master thread we create
a new ThreadPool.

This needs a test and some more careful thought.  WIP

* add test for threaded-thread computation

* Track and clean up inactive thread pools

* add lock around pool manipulation