``dd.map_partitions`` works with scalar outputs (#1515)
Previously if a user defined function returned a scalar, the output of
`dd.map_partitions` would be a `Scalar`, even if the input type was
a series. This fixes that so that scalar returning functions return a
series, unless all inputs are scalars (in which case the output is a
scalar as before). Fixes
http://stackoverflow.com/questions/39215617/what-is-map-partitions-doing.