Avoid pickle when tokenizing __main__ functions
Previously two functions defined in main with the same name would
tokenize to the same value.  This is because `pickle` would return
the same value in both cases without raising an exception (as might be
expected)

This now checks the result for the bytestring b`__main__` and if present
continues on to use cloudpickle.

Fixes https://github.com/dask/distributed/issues/487