Only fuse simple indexing with getarray backends
Certain numpy-array-like-things may not support all of the indexing
methods that numpy uses. When fusing getitem/getarray calls together, we
shouldn't pass "fancy indexing" requests to these backends. The easiest
way to do this is to not propogate fusing of these types back inside
calls to `getarray`.

This PR includes the following changes:
- Don't fuse indexing using lists or arrays inside calls to `getarray`.
- Make `from_array` use `getitem` when used with numpy types that return
  instances of `ndarray` when indexed. This is a small optimization to
  make use of `np.memmap` and `np.ndarray` still as fast as before.