Bring back forked workers
Recently we switched to having Nannys create new workers with subprocess
rather than fork.  This allowed us to spawn workers in different python
environments.

However, this adds some cruft, both recent issues with cleanly stopping
workers and a non-trivial delay in startup time (multiple seconds).

Now we support both fork and subprocess, using the latter only when the
requested environment is not the same environment as the Nanny's