Sample from `read_bytes` ends on a delimiter
If a delimiter is provided to `read_bytes`, the sample will end with the
delimiter.  This was specified in the docstring, but wasn't actually
implemented, since the delimiter wasn't forwarded. The `sample` kwarg
was also dropped in a few places throughout.

This PR fixes these issues, and removes the need to adjust the end of
the sample in ``dd.read_csv`` added in #1495. Pandas won't error a
numeric field is errored, which led to the failure in #1532. A test has
been added for this, and confirmed to fix on the provided csv.

Fixes #1532.