Tokenize mmap works without filename (#1570)
Due to a bug in numpy, ufuncs performed on `memmap`'s return an
instance of `memmap`, even though the data is already in memory. In
these cases `memmap.filename` is `None`, which causes `tokenize` to
error. We now treat `memmap` files where `filename is None` the same as
an in-memory file.

Fixes #1562.