Groupby works with multiprocessing (#1625)
Previously this would fail, as the partd was buffered, and wouldn't
write to disk some/all of the data. Since we still want to use the
buffer with the threaded scheduler, we solve this by switching behavior
if the partd creation function is serialized. If the function isn't
serialized, a buffer is used, if it is serialized then a file based
partd is used.

Fixes #1587.