Remove use of multiprocessing.Manager (#1652)
Using a Manager process makes multiprocessing.get fragile on Windows
as concurrent unpickling of Numpy objects may result in sporadic errors
(the numpy module being seen uninitialized by one of the manager threads).

Instead of queueing explicitly in execute_task(), simply return the
result and queue it in the parent.