Add rename support for multi-level columns
Fixes #1694.

Also found a bug in the method versions of operators when applied to
Series with axis=1. `map_partitions` was automatically repartitioning
this, resulting in an incorrect answer. However, the old way of renaming
columns was actually creating new columns filled with `NaN`. Since the
data it was being tried had no matches between columns and index, the
output was also all `NaN`, masking the issue. This changes the test so
it would have caught the issue, and fixes the implementation of the
operator.