Add rename support for multi-level columns (#1712)
* Add rename support for multi-level columns

Fixes #1694.

Also found a bug in the method versions of operators when applied to
Series with axis=1. `map_partitions` was automatically repartitioning
this, resulting in an incorrect answer. However, the old way of renaming
columns was actually creating new columns filled with `NaN`. Since the
data it was being with tried had no matches between columns and index, the
output was also all `NaN`, masking the issue. This changes the test so
it would have caught the issue, and fixes the implementation of the
operator.

Also fixed a few bugs in `_rename` and `map_partitions`:

- `_rename` now returns `Index` when passed an `Index`. Previously this
  would return a `Series`.

- `map_partitions` now returns partitions with the same columns in all
  occasions. Previously partitions with non-matching columns could sneak
  through, which would cause issues later.