Timeout and check on write (#653)
* Timeout and check on write

The core.write coroutine waits on IOStream.write futures.  However, there are
some cases where IOStream.write futures don't complete.  Notably if something
else writes to the same stream before we're done waiting then the first future
is forgotten and hangs forever.

    If `write` is called again before that `.Future` has resolved, the
    previous future will be orphaned and will never resolve.

This solution wraps the wait in a timeout and checks the stream's
write_buffer as a secondary check for completion.  There are probably better
solutions but this seems to work fine.

* Remove timeout trick from batched

This was recently moved to core.write