ProgressBar exits on error with get_sync
Due to some convoluted logic in when and how the `finish` callbacks were
run in `get_async`, the `finish` method wouldn't be called if
`raise_on_exception` was `True`. Since `get_sync` sets this as `True`,
the `ProgressBar` would never exit if exceptions occurred.

This simplifies the logic, and ensures that `finish` is always called on
exceptions.

Fixes #1791.