Fix `drop_duplicates` with `split_out`
The `split_out` kwarg was added to `aca` in 303d038. This allows for
outputs to be split by hashing the rows, improving efficiency for large
outputs. However, this defaulted to hashing all columns in the output
dataframe, which didn't play well with the `subset` kwarg to
`drop_duplicates`. This was resulting in not all duplicate rows being
dropped, as they'd hash to different partitions since the hashed columns
weren't being subset.

To fix this, we replace the `split_index` kwarg with `split_out_setup`,
which optionally takes a function to apply to each chunk before being
hashed. `split_out_setup_kwargs` is also available to pass keywords to
this function. `drop_duplicates` was modified to adjust for this change.