Filter out empty frames in `_concat`.
We do this because pandas frequently has inconsistent dtypes in
operations with empty dataframes. We may want to move this into
`methods.concat`, but at the same time that needs to be able to use
empty frames to ensure partitions match the desired schema. This all is
pretty mucky stuff, hopefully can find a cleaner way.