Use `delayed(x).compute()` instead of `compute(*x)`
Previously `compute(*args)` is used in several places to compute a list
of `Delayed` values - usually for I/O purposes. This uses the generic
`compute` function from `dask.base`. This is a generic method that
has no knowledge of `Delayed`, and instead works with the `dask`
attributes on objects. Because of this, it does unnecessary merging of
dictionaries that can be skipped by using `delayed(args).compute()`
instead. For large numbers of `Delayed` objects, this can be much
faster, and uses less memory.

Fixes #1888.