Optimize `compute` to avoid unnecessary graph merging
Previously `dask.base.compute` would only look for the `dask` attribute
on dask objects. This was general, but could cause unnecessary copying
in the case of many `dask.delayed.Delayed` values. We now first check
for the existence of the `_dasks` attribute (as found on `Delayed`
values), and fallback to the `dask` attribute if not found. For a
pathological case, this results in a 2500% memory reduction.