Use `delayed(x).compute()` instead of `compute(*x)` (#1894)
* Use `delayed(x).compute()` instead of `compute(*x)`

Previously `compute(*args)` is used in several places to compute a list
of `Delayed` values - usually for I/O purposes. This uses the generic
`compute` function from `dask.base`. This is a generic method that
has no knowledge of `Delayed`, and instead works with the `dask`
attributes on objects. Because of this, it does unnecessary merging of
dictionaries that can be skipped by using `delayed(args).compute()`
instead. For large numbers of `Delayed` objects, this can be much
faster, and uses less memory.

Fixes #1888.

* Optimize `compute` to avoid unnecessary graph merging

Previously `dask.base.compute` would only look for the `dask` attribute
on dask objects. This was general, but could cause unnecessary copying
in the case of many `dask.delayed.Delayed` values. We now first check
for the existence of the `_dasks` attribute (as found on `Delayed`
values), and fallback to the `dask` attribute if not found. For a
pathological case, this results in a 2500% memory reduction.