quote handles lists, faster
Previously `quote` would only quote tasks, and would make a copy of the
tuple, as well as recurse into its arguments. Now we wrap literals in a
small class, and then return a task to unbox them. This is more
efficient, as it removes the need to copy and is O(1) (instead of
recursing through the collection).