Add `traverse` kwarg to `delayed` (#1899)
* Add `traverse` kwarg to `delayed`

By default dask traverses builtin python collections looking for dask
objects passed to ``delayed``. For large collections this can be
expensive. This PR adds a keyword ``traverse`` to ``delayed``, to tell
dask to avoid doing this traversal.

Fixes #1884.

* quote handles lists, faster

Previously `quote` would only quote tasks, and would make a copy of the
tuple, as well as recurse into its arguments. Now we wrap literals in a
small class, and then return a task to unbox them. This is more
efficient, as it removes the need to copy and is O(1) (instead of
recursing through the collection).

* Simplify literal slightly

Use `__call__` instead of a separate function.